Moved DataObject to robot
