+++ ROVU/Core/src/robot/DataObject.java
@@ -1,4 +1,4 @@
-package controller;
+package robot;
 
 import java.util.List;
